requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
httpx>=0.27.0
websockets>=12.0
pandas>=2.2.0
numpy>=1.26.0
sqlalchemy>=2.0.29
sqlmodel>=0.0.16
asyncpg>=0.29.0
alembic>=1.13.0
redis>=5.0.0
apscheduler>=3.10.4
python-dotenv>=1.0.1
structlog>=24.1.0
py-clob-client>=0.17.0
streamlit>=1.32.0
plotly>=5.20.0
textblob>=0.18.0
